requirements.txt
alembic==1.13.1
asn1crypto==1.5.1
Babel==2.15.0
bcrypt==4.1.3
bleach==6.1.0
blinker==1.8.2
cachelib==0.9.0
cbor2==5.6.3
cffi==1.16.0
click==8.1.7
cryptography==42.0.7
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
flask-babel==4.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
flask-mailman==1.1.0
Flask-Principal==0.4.0
Flask-Security-Too==5.4.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.0
passlib==1.7.4
phonenumberslite==8.13.37
pluggy==1.5.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pypng==0.20220715.0
pytest==8.2.1
pytz==2024.1
qrcode==7.4.2
six==1.16.0
SQLAlchemy==2.0.30
SQLAlchemy-Utils==0.41.2
typing_extensions==4.12.0
webauthn==2.1.0
webencodings==0.5.1
Werkzeug==3.0.3
WTForms==3.1.2
